backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.8.0
asyncpg==0.29.0
psycopg[binary,pool]==3.1.15
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.6
litellm==1.35.12
tiktoken==0.5.2
google-generativeai==0.8.6